pyproject.toml
[project]
name = "langchain-docs-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-community>=0.3.14",
    "langchain-huggingface>=0.1.2",
    "langchain-ollama>=0.2.2",
    "langchain-pinecone>=0.2.2",
    "langchainhub>=0.1.21",
    "pinecone>=5.4.2",
    "requests==2.27.1",
    "ruff>=0.9.2",
    "streamlit-chat>=0.1.1",
    "streamlit>=1.41.1",
    "tiktoken>=0.8.0",
    "torch==2.2.2",
    "tqdm>=4.67.1",
    "uvicorn>=0.34.0",
    "python-dotenv>=1.0.1",
    "ipython>=8.31.0",
    "bs4>=0.0.2",
    "watchdog>=6.0.0",
    "firecrawl-py>=1.10.2",
    "unstructured>=0.16.16",
]
